feat:: removed some not needed folder
